requirements.txt
picamera
opencv-python o
pencv-python-headless 
telepot
pyttsx3
RPi.GPIO
